netlink: improve documentation, add additional comments
Document that deadlines only work on Go 1.12+.

Internally, document that we use non-blocking file descriptors on
Go 1.12+, and that our call to os.NewFile registers these with the
runtime poller.

Document that for Go <1.12, we maintain compatibility with the old
behavior, namely the use of blocking file descriptors.

Drop TODO comment about Read and Write methods on rawConn for Go <1.12,
since those methods will not be supported.